feat: add general IR metrics module and update rubric to 0-1 float scale
Step 1: Add metrics module with domain-agnostic IR ranking functions
(hit_rate, MRR, recall_at_k, precision_at_k, average_precision) and
content-based overlap matching (text_overlap_ratio, is_content_match,
content_reciprocal_rank) with full test coverage (44 tests).

Step 2: Update judge rubric from 1-5 integer scale with 4 dimensions
(correctness, grounding, completeness, clarity) to 0-1 float scale
with 3 dimensions (correctness, faithfulness, relevance). Updated
scoring, prompts, report formatting, and all related tests.

+++ src/judge/scoring.py
@@ -8,8 +8,8 @@ from typing import Any, Dict, Optional
 class JudgeResult:
     trace_id: str
     rubric_name: str
-    scores: Dict[str, int]
-    overall: Optional[int] = None
+    scores: Dict[str, float]
+    overall: Optional[float] = None
     rationale: Optional[str] = None
     raw: Optional[Dict[str, Any]] = None
 
@@ -26,16 +26,16 @@ class JudgeResult:
 
 def normalize_judge_output(trace_id: str, rubric_name: str, output: Dict[str, Any]) -> JudgeResult:
     scores = output.get("scores") or {}
-    norm_scores: Dict[str, int] = {}
+    norm_scores: Dict[str, float] = {}
     for k, v in scores.items():
         try:
-            norm_scores[str(k)] = int(v)
+            norm_scores[str(k)] = float(v)
         except Exception:
             continue
 
     overall = output.get("overall")
     try:
-        overall = int(overall) if overall is not None else None
+        overall = float(overall) if overall is not None else None
     except Exception:
         overall = None
